requirements.txt
pandas>=2.0.0
selenium>=4.0.0
python-docx>=1.0.0
openpyxl>=3.0.0
webdriver-manager>=4.0.0